Fix incorrect binlog file opened when building solution
The original code was not handling multiple binlog files that can
be created when building the solution. This could result in the wrong
one being opened from the Build Logging window. To fix this all
build started events are captured and associated with any MSBuild
target that is run whilst they are active. This may result in multiple
binlogs being associated with a single MSBuild target initially.
When the build finished event is fired the binlog is processed to
see what projects are defined in the binlog. This is then used to
associate the binlog with the MSBuild target displayed in the Build
Logging window.