Fix not being able to open Clean binlog on re-building solution
The built-in structured build output support in the IDE will delete
the Clean binlogs for projects on rebuilding the solution. It also
deletes the old Build binlogs if a build is run again for the
solution. To avoid the binlogs being unavailable the binlogs files
are copied to a backup file when the solution build session is
finished.

Note that Cleaning a single project does not generate a binlog.
Building a single project will generate a binlog and create a
build session so when the build session finishes a backup copy
will be made of the binlog.