Support capturing a binlog for solution NuGet restore
If a solution contains more than 10 projects the NuGet restore is
done using MSBuild directly instead of the build host. To capture
a binlog the MSBuildProcessService has been modified to allow the
process start to be intercepted. The command line arguments passed
to MSBuild are changed so a binlog is generated using the -bl
parameter. This allows the GenerateRestoreGraphFile MSBuild target
to generate a binlog which can be opened from the Build Logging
window. The normal MSBuild text output is not currently available
from the Build Logging window. It is available from the NuGet
addin's Package Console.

This requires a custom MonoDevelop build.