Capture msbuild console output when using MSBuildProcessService
Running a NuGet restore for a solution that contains more than 10
PackageReference projects uses the MSBuildProcessService. In order to
capture the msbuild stdout and stderr output the TextWriter's passed
to the ProcessService need to be replaced with custom ones that
can be used to capture the output and save it to a file. The custom
TextWriter's chain the existing TextWriter's which were passed to
MSBuildProcessService so the output is still sent to the original
writers.