Fix no target/task summary information shown after building solution
Building a solution, or building a project that referenced another
project, the Task Summary and Target Summary tabs would not show
any information. This was because the project-system-tools code
does not support multiple binlog projects that do not have a parent.
msbuild and Visual Studio on Windows do not need this support since
they always have a single project with no parent project in the
binlog. Visual Studio for Mac may produce multiple projects in the
binlog without any parent project. The project-system-tools code
has been forked and modified to support this.