Allow sorting of columns in Build Logging pad
Unfortunately after sorting, when a new item is added, the text
in all the columns is not displayed. Possibly an XWT bug. A possible
workaround would be to recreate the list view and list store. This
workaround was added to the NuGet manage packages dialog which had
a similar problem.