Workaround Save As dialog being displayed twice
File - Save As shows the Save As dialog twice. This seems to be a
problemg in VS Mac's BuildOutputViewContent. It's SaveAs method
shows the save as dialog but the FileDocumentController's SaveAs
already does this.

A workaround has been added here. The DocumentControllerExtension
overrides the OnSave method which prevents the default logic in
the BuildOutputViewContent being run. This new OnSave method copies
the original binlog file to the new location.

Fixes #2