Support MSBuild 16.10 binlogs
The Properties property on the event args for the
project started event is null. Use the project file
name from the event args instead. This fixes the
Open Binary Log menu item not being available when
right clicking a row in the Build Logging window
since no project file was mapped to a binlog.