Remove insertbefore from project extensions
insertbefore="InitialStep" does not add the extension as the first
in the chain, but instead adds it at the end since VS Mac seems to
revert the order. Adding the extension without the insertbefore
will add the extension further down the chain but possibly not the
last since it depends on the order that addins are loaded.